pyproject.toml
[project]
name = "deng"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "google-cloud-storage>=3.10.1",
    "jupyter>=1.1.1",
    "pandas>=3.0.1",
    "swissparlpy>=1.0.0",
]
